requirements.txt
Flask==1.1.2
line-bot-sdk==1.17.0
gunicorn==20.0.4
pandas==1.2.3
xlrd==2.0.1
openpyxl==3.0.7
pymessenger==0.0.7.0